// src/app/shared/utils/enum.ts
export const enum STATUS {
  EXECUTADO = "EXECUTADO",
  PROGRAMADO = "PROGRAMADO",
  ANDAMENTO = "ANDAMENTO",
}

export const enum EMBARGOES {
  LIBERADO = "LIBERADO",
  PROJETO = "PROJETO",
  FUNDIARIO = "FUNDIARIO",
  ARQUEOLOGIA = "ARQUEOLOGIA"
}

export const enum STATUS_EMPLOYEE {
  ATIVO = "ATIVO",
  TREINAMENTO = "TREINAMENTO",
  AFASTADO = "AFASTADO"
}

export const enum LEADERSHIP {
  SIM = "SIM",
  NAO = "NAO"
}

export const enum STATUS_EQUIPMENT {
  ATIVO = "ATIVO",
  MANUTENCAO = "MANUTENCAO",
}

export const enum WEATHER {
  BOM = "BOM",
  CHUVA = "CHUVA",
  IMPRATICAVEL = "IMPRATICAVEL"
}

export const enum STAGE {
  PRELIMINAR = "PRELIMINAR",
  CIVIL = "CIVIL",
  MONTAGEM = "MONTAGEM",
  LANCAMENTO = "LANCAMENTO",
  COMISSIONAMENTO = "COMISSIONAMENTO"
}

export const enum UNITY {
  TORRE = "TORRE",
  KM = "KM",
  UNIDADE = "UNIDADE",
}